Update Summary file structure
