+++ context/app/static/js/components/Detail/Summary/index.js
@@ -1,0 +1,3 @@
+import Summary from './Summary';
+
+export default Summary;
